+++ internal/cmd/docker/dockercompose/docker-compose.go
@@ -16,10 +16,11 @@ var dockerComposeExample string
 
 func NewCmd() *cobra.Command {
 	return &cobra.Command{
-		Use:     "docker-compose [flags] [docker compose command] -- [docker compose flags]",
+		Use:     "compose [flags] [compose command] -- [compose flags]",
+		GroupID: "core",
 		Example: dockerComposeExample,
-		Aliases: []string{"dc"},
-		Short:   "Interact directly with docker compose",
+		Aliases: []string{"dc", "compose"},
+		Short:   "Interact directly with the docker compose environment",
 		Long:    cli.RenderMarkdown(dockerComposeLong),
 		Run: func(cmd *cobra.Command, args []string) {
 			dev := mwdd.DefaultForUser()