+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Each tagged release MUST have a section 2 heading starting at the time of releas
 
 - Minor internal package adjustments
 - Improved error message for `docker <service> expose` commands when container is not running
+- Improved layout of grouped sub commands
 
 ## v0.24.0